Update main README with user-friendly directions
The ceph.io website as of now is not friendly towards those who would like to fix or at least to report mistakes on the website. This repo does not make it clear that it is for the new website under development and this creates confusion. See issue #3 for an example. It is safe to assume that the majority of people who would have otherwise reported or even fixed mistakes give up early in the process. This update provides direction for reports related to the current website and makes a statement about this repository's purpose as a home for the new website.