ceph.io/developers: rewriting marketing copy 1
This is the first PR in a series of PRs that will
remove the marketing copy on the Developers page of the
ceph.io website. This PR:

- Links to the part of the Developer Guide that
  explains how to make a PR to the Ceph Github
  repo.
- Provides a link to the Github repo.
- Provides a link to the Ceph Issue Tracker (the
  Redmine instance).

This provides the three basic tools that a potential
contributor will need to make a contribution to the
project.

Note: there will be more PRs to this file in the near
future after I consult with Sage. If you are not Sage
and you are reading this, don't think that this is the
only change that is being made to this page.

Signed-off-by: Zac Dover <zac.dover@gmail.com>