ceph.io/discover: change header text
This PR changes the header text from "Achieve
excellent performance, reliability and scalabitily
with Ceph." to "Ceph. The future of storage."

I'm putting this in a PR now so that I can get it out
of the way. Fixing this is going to be a death of a
thousand cuts.

Signed-off-by: Zac Dover <zac.dover@gmail.com>