ceph.io/contribute: add contribute video
This PR adds the Youtube video that explains how
to contribute to the Ceph community. I've put it here
in iframe html tags, but for some reason they don't work
within the website.

Maybe we can boot this to someone with greater understanding
of embedding video in this IA.

Update:
Changed the formatting of the video link from iframe to
YouTube embed shortcode.

Signed-off-by: Zac Dover <zac.dover@gmail.com>