doc/ceph.io: discover -> technology (1 of x)
This is the first in a series of PRs that will
rewrite the "Discover -> Technology" page. This
page adds meaningful text and links to documentation
that more completely explains the components referred
to on this page.

Lorem ipsum text remains present here. When the technical
content added in this PR is approved, I'll ask Pete to remove
the lorem ipsum text.

Signed-off-by: Zac Dover <zac.dover@gmail.com>