Improve Cephalocon 2024 page
Make as many changes to the Cephalocon 2024 page as it possible to make
of the ones requested by Bridgette Waldren in May 2024. Unfortunately,
Eleventy (the software that we use to generate the ceph.io website)
doesn't permit any date format other than YYYY-MM-DD, so Bridgette's
requests for the dates to be in DD-MM-YYYY format run up against the (in
my opinion, unreasonably hard-nosed and inflexible) prohibitions of
Eleventy.

Signed-off-by: Zac Dover <zac.dover@gmail.com>